+++ vast_csi/builders/block.py
@@ -36,6 +36,8 @@ class BlockProvisionBase(BaseVolumeBuilder):
     vip_pool_name: Optional[str] = None
     vip_pool_fqdn: Optional[str] = None
     vip_pool_fqdn_random_prefix: Optional[bool] = None
+    qos_policy: Optional[str] = None
+    qos_policy_id: Optional[int] = None
     blocking_clones: Optional[bool] = None
     capacity_range: Optional[int] = None
     pvc_name: Optional[str] = None
@@ -67,6 +69,11 @@ class BlockProvisionBase(BaseVolumeBuilder):
         cls._validate_mount_src(vip_pool_name, vip_pool_fqdn, conf.use_local_ip_for_mount)
         cluster_name = parameters.get("cluster_name")
         blocking_clones = cls._get_bool_param(parameters, "blocking_clones")
+        qos_policy = parameters.get("qos_policy")
+        if "qos_policy_id" in parameters:
+            qos_policy_id = int(parameters.get("qos_policy_id"))
+        else:
+            qos_policy_id = None
 
         return cls(
             vms_session=vms_session,
@@ -83,6 +90,8 @@ class BlockProvisionBase(BaseVolumeBuilder):
             host_encryption=host_encryption,
             vip_pool_fqdn_random_prefix=vip_pool_fqdn_random_prefix,
             cluster_name=cluster_name,
+            qos_policy_id=qos_policy_id,
+            qos_policy=qos_policy,
             volume_content_source=volume_content_source,
             blocking_clones=blocking_clones,
             **metadata,
@@ -145,11 +154,20 @@ class EmptyBlockVolumeBuilder(BlockProvisionBase):
             subsystem=self.subsystem,
             tenant_name=self.tenant_name,
         )
-        volume = self.vms_session.volumes.ensure(
+        qos_policy_id = self.qos_policy_id
+        if self.qos_policy:
+            qos_policy_id = self.vms_session.quospolicies.one(
+                name=self.qos_policy,
+                fail_if_missing=True,
+            ).id
+        volume_data = dict(
             name=volume_name,
             view_id=view.id,
             size=requested_capacity,
         )
+        if qos_policy_id:
+            volume_data["qos_policy_id"] = qos_policy_id
+        volume = self.vms_session.volumes.ensure(**volume_data)
         volume_context.update(
             nguid=volume.nguid,
             volume_id=str(volume.id),
@@ -208,6 +226,17 @@ class BlockVolumeFromVolumeBuilder(BlockProvisionBase):
                 target_subsystem_id=destination_view.id,
                 target_volume_path=volume_name,
             )
+            qos_policy_id = self.qos_policy_id
+            if self.qos_policy:
+                qos_policy_id = self.vms_session.quospolicies.one(
+                    name=self.qos_policy,
+                    fail_if_missing=True,
+                ).id
+            if qos_policy_id:
+                self.vms_session.volumes.update(
+                    destination_volume.id,
+                    qos_policy_id=qos_policy_id,
+                )
 
         if self.blocking_clones:
             # Wait for the clone to be ready
@@ -261,6 +290,17 @@ class BlockVolumeFromSnapshotBuilder(BlockProvisionBase):
                 target_subsystem_id=destination_view.id,
                 target_volume_path=volume_name,
             )
+            qos_policy_id = self.qos_policy_id
+            if self.qos_policy:
+                qos_policy_id = self.vms_session.quospolicies.one(
+                    name=self.qos_policy,
+                    fail_if_missing=True,
+                ).id
+            if qos_policy_id:
+                self.vms_session.volumes.update(
+                    destination_volume.id,
+                    qos_policy_id=qos_policy_id,
+                )
 
         if self.blocking_clones:
             # Wait for the clone to be ready